requirements.txt
dill==0.3.7
gymnasium==0.28.1
numpy==1.26.4
pygame==2.5.2
ray==2.34.0
ray[rllib]==2.34.0
torch==2.3.0
hydra-core==1.3.2
pettingzoo==1.24.2